Restyling finished for search bar
