+++ frontend/src/components/map/Map.css
@@ -1,7 +1,7 @@
 /* Map */
 .map-container { 
     position: relative;
-    width: 100vw;
+    max-width: 100vw;
     /* If navbar height changes, change this */
     height: calc(100vh - 96px);
     z-index: 0;
