+++ frontend/src/components/viewToggler/ViewToggler.css
@@ -1,14 +1,3 @@
-.options-bar {
-    display: block;
-}
-
-@media (min-width: 576px) {
-    .options-bar {
-        display: flex;
-        flex-wrap: wrap;
-        justify-content: space-between;
-    }
-}
 
 .view-btn {
     z-index: 1;
@@ -31,12 +20,4 @@
 
 .view-btn i {
     padding: 5px;
-}
-
-.options-bar .select-container {
-    padding-bottom: 0.5rem;    
-}
-
-.options-bar select {
-    width: 100%;
 }
